Add files via upload
I will do more of the project in the weekend, but i want to have the main html finished by the end of the week. I will add scss and all the format by the end. I have some impediments:
back-ground image does not show on product page, I have tried everything.
Search bar does not align to the right of the bage. 
The carousel on the main page does not center and the arrows don't show at the extremities of the page.

When I will get feedback from you I will get and work on it. Thanks